requirements.txt
# Smart Parking Management System - Dependencies
pandas
matplotlib
numpy
Pillow
qrcode[pil]
reportlab
ttkthemes
faker
opencv-python
pytesseract
imutils
pyzbar
django
djangorestframework